MorphoSyntaxTrening/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
starlette==0.46.2
jinja2==3.1.6
python-multipart==0.0.26
sqlalchemy==2.0.49
alembic==1.18.4
itsdangerous>=2.1.2
passlib[bcrypt]>=1.7.4
bcrypt>=3.1.0,<4.0
pymorphy3>=2.0.0
pdfminer.six>=20221105
